feat: implement docker-compose solution, refactor src structure
