+++ Dockerfile
@@ -1,24 +1,13 @@
-FROM node:16 AS builder
+FROM node:20-alpine
 
-# Create app directory
 WORKDIR /app
 
-# A wildcard is used to ensure both package.json AND package-lock.json are copied
 COPY package*.json ./
+
+RUN npm install && npm cache clean --force
+
 COPY prisma ./prisma/
 
-# Install app dependencies
-RUN npm install
+RUN npx prisma generate
 
-COPY . .
-
-RUN npm run build
-
-FROM node:16
-
-COPY --from=builder /app/node_modules ./node_modules
-COPY --from=builder /app/package*.json ./
-COPY --from=builder /app/dist ./dist
-
-EXPOSE 3000
-CMD [ "npm", "run", "start:prod" ]
+CMD ["sh", "-c", "npx prisma migrate deploy && npm run start:dev"]
