+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AlbumsModule } from './albums/albums.module';
+import { AlbumsModule } from './modules/albums/albums.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ArtistsModule } from './artists/artists.module';
-import { FavoritesModule } from './favorites/favorites.module';
-import { TracksModule } from './tracks/tracks.module';
-import { UsersModule } from './users/users.module';
+import { ArtistsModule } from './modules/artists/artists.module';
+import { FavoritesModule } from './modules/favorites/favorites.module';
+import { TracksModule } from './modules/tracks/tracks.module';
+import { UsersModule } from './modules/users/users.module';
 
 @Module({
   imports: [
