+++ src/modules/albums/albums.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/modules/prisma/prisma.module';
 import { AlbumsController } from './albums.controller';
 import { AlbumsService } from './albums.service';
