+++ src/modules/artists/artists.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/modules/prisma/prisma.module';
 import { ArtistsController } from './artists.controller';
 import { ArtistsService } from './artists.service';
