+++ src/modules/artists/dto/update-artist.dto.ts
