+++ src/modules/artists/interfaces/artists.interface.ts
