+++ src/modules/favorites/favorites.controller.ts
