+++ src/modules/favorites/interfaces/favorite.interface.ts
