+++ src/modules/tracks/tracks.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/modules/prisma/prisma.module';
 import { TracksController } from './tracks.controller';
 import { TracksService } from './tracks.service';
